requirements.txt
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.0
openpyxl>=3.1.0
numpy>=1.24.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
